vaana_app/requirements.txt
asgiref==3.3.4
Django==3.2.1
django-cors-middleware==1.5.0
djoser==2.1.0
drf-yasg==1.20.0
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
graphviz==0.16
psycopg2==2.8.6
pytz==2021.1
six==1.14.0
sqlparse==0.4.1
gunicorn
django-seed
requests
stripe
django-storages
boto3
chatterbot-corpus==1.2.0
requests
shippo
django-paypal
braintree
# ChatterBot==1.0.4
# chatterbot-corpus==1.2.0
google-api-core==2.0.1
google-api-python-client==2.19.1
google-auth==2.0.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
facebook-sdk==3.1.0
